simplify orm repository
