+++ src/Collection/Doctrine/ORM/Repository/IsCollection.php
@@ -3,7 +3,7 @@
 namespace Zenstruck\Collection\Doctrine\ORM\Repository;
 
 use Zenstruck\Collection;
-use Zenstruck\Collection\Doctrine\ORM\ObjectRepository;
+use Zenstruck\Collection\Doctrine\ORM\EntityRepository;
 
 /**
  * Enables your repository to implement {@see Collection} (and use
@@ -20,8 +20,8 @@ trait IsCollection
      */
     final public function take(int $limit, int $offset = 0): Collection
     {
-        if (!$this instanceof ObjectRepository) {
-            throw new \BadMethodCallException(\sprintf('"%s" can only be used on instances of "%s".', __TRAIT__, ObjectRepository::class));
+        if (!$this instanceof EntityRepository) {
+            throw new \BadMethodCallException(\sprintf('"%s" can only be used on instances of "%s".', __TRAIT__, EntityRepository::class));
         }
 
         return static::createResult($this->qb())->take($limit, $offset);
