+++ src/Collection/Doctrine/ORM/Repository/IsFilterable.php
@@ -3,7 +3,7 @@
 namespace Zenstruck\Collection\Doctrine\ORM\Repository;
 
 use Doctrine\ORM\QueryBuilder;
-use Zenstruck\Collection\Doctrine\ORM\ObjectRepository;
+use Zenstruck\Collection\Doctrine\ORM\EntityRepository;
 use Zenstruck\Collection\Doctrine\ORM\Result;
 use Zenstruck\Collection\Doctrine\ORM\Specification\ORMContext;
 use Zenstruck\Collection\Filterable;
@@ -26,8 +26,8 @@ trait IsFilterable
      */
     final public function filter(mixed $specification): Result
     {
-        if (!$this instanceof ObjectRepository) {
-            throw new \BadMethodCallException(\sprintf('"%s" can only be used on instances of "%s".', __TRAIT__, ObjectRepository::class));
+        if (!$this instanceof EntityRepository) {
+            throw new \BadMethodCallException(\sprintf('"%s" can only be used on instances of "%s".', __TRAIT__, EntityRepository::class));
         }
 
         if (\is_array($specification) && !array_is_list($specification)) {
@@ -43,8 +43,8 @@ trait IsFilterable
      */
     final public function get(mixed $specification): object
     {
-        if (!$this instanceof ObjectRepository) {
-            throw new \BadMethodCallException(\sprintf('"%s" can only be used on instances of "%s".', __TRAIT__, ObjectRepository::class));
+        if (!$this instanceof EntityRepository) {
+            throw new \BadMethodCallException(\sprintf('"%s" can only be used on instances of "%s".', __TRAIT__, EntityRepository::class));
         }
 
         if (\is_scalar($specification) || (\is_array($specification) && !array_is_list($specification))) {
