+++ stubs/ORMRepository.php
@@ -1,17 +1,17 @@
 <?php
 
 use Zenstruck\Collection;
+use Zenstruck\Collection\Doctrine\ORM\EntityRepository;
 use Zenstruck\Collection\Doctrine\ORM\Repository\IsFilterable;
 use Zenstruck\Collection\Doctrine\ORM\Repository\IsCollection;
-use Zenstruck\Collection\Doctrine\ORMRepository as BaseORMRepository;
 use Zenstruck\Collection\Paginatable;
 
 /**
  * @template V of object
- * @extends BaseORMRepository<V>
+ * @extends EntityRepository<V>
  * @implements Collection<int,V>
  */
-abstract class ORMRepository extends BaseORMRepository implements Collection
+abstract class ORMRepository extends EntityRepository implements Collection
 {
     /** @use IsFilterable<V> */
     use IsFilterable;
