+++ tests/Doctrine/ORM/Fixture/KitchenSinkRepository.php
@@ -2,12 +2,10 @@
 
 namespace Zenstruck\Collection\Tests\Doctrine\ORM\Fixture;
 
+use Doctrine\ORM\EntityManagerInterface;
 use Zenstruck\Collection;
-use Zenstruck\Collection\Doctrine\ORM\Repository\AsEntityRepository;
-use Zenstruck\Collection\Doctrine\ORM\Repository\AsService;
 use Zenstruck\Collection\Doctrine\ORM\Repository\IsCollection;
 use Zenstruck\Collection\Doctrine\ORM\Repository\IsFilterable;
-use Zenstruck\Collection\Doctrine\ORMRepository;
 use Zenstruck\Collection\Filterable;
 use Zenstruck\Collection\Paginatable;
 use Zenstruck\Collection\Tests\Doctrine\Fixture\Entity;
@@ -15,12 +13,12 @@ use Zenstruck\Collection\Tests\Doctrine\Fixture\Entity;
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
  */
-final class KitchenSinkRepository extends ORMRepository implements Filterable, Collection
+final class KitchenSinkRepository extends Collection\Doctrine\ORM\EntityRepository implements Filterable, Collection
 {
-    use AsEntityRepository, AsService, IsCollection, IsFilterable, Paginatable;
+    use IsCollection, IsFilterable, Paginatable;
 
-    public function getClassName(): string
+    public function __construct(EntityManagerInterface $em)
     {
-        return Entity::class;
+        parent::__construct($em, $em->getClassMetadata(Entity::class));
     }
 }
